ENH: Port ugrid-tools #466
- Functionality from the ugrid-tools library has been merged into OCGIS.
  This includes interior splitting, node reduction, and writing of the
  ESMF Unstructured format.
- Ashynchronous NetCDF writing was added. This is supported by new
  versions of netCDF4-python. #470
- Full support for string variable writing also added. #379

+++ src/ocgis/test/test_ocgis/test_collection/test_field.py
@@ -10,7 +10,7 @@ from shapely.geometry import Point
 from shapely.geometry import box
 from shapely.geometry.base import BaseGeometry
 
-from ocgis import RequestDataset, vm, DimensionMap, env
+from ocgis import RequestDataset, vm, DimensionMap
 from ocgis import constants
 from ocgis.base import get_variable_names
 from ocgis.collection.field import Field, get_name_mapping
@@ -20,13 +20,12 @@ from ocgis.conv.nc import NcConverter
 from ocgis.driver.csv_ import DriverCSV
 from ocgis.driver.nc import DriverNetcdf
 from ocgis.driver.vector import DriverVector
-from ocgis.ops.core import OcgOperations
 from ocgis.spatial.geom_cabinet import GeomCabinetIterator
 from ocgis.spatial.grid import Grid, create_grid_mask_variable
 from ocgis.test.base import attr, AbstractTestInterface
 from ocgis.util.helpers import reduce_multiply
 from ocgis.variable.base import Variable
-from ocgis.variable.crs import CoordinateReferenceSystem, WGS84, Spherical, CRS
+from ocgis.variable.crs import CoordinateReferenceSystem, WGS84, Spherical
 from ocgis.variable.dimension import Dimension
 from ocgis.variable.geom import GeometryVariable
 from ocgis.variable.temporal import TemporalVariable
@@ -261,7 +260,7 @@ class TestField(AbstractTestInterface):
         gci = GeomCabinetIterator(path=self.path_state_boundaries)
         actual = Field.from_records(gci, data_model='NETCDF3_CLASSIC')
         desired = {'UGID': np.int32,
-                   'ID': np.float32}
+                   'ID': np.int32}
         for v in desired.keys():
             self.assertEqual(actual[v].get_value().dtype, desired[v])
